# src/cleansia_customer_android/openapi/customer-api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cleansia.Customer.API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Order/Lookup": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "Cleansia.Web.Customer.Controllers.OrderController.LookupOrder (Cleansia.Web.Customer)",
        "parameters": [
          {
            "name": "orderNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LookupOrder_Response"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupOrder_Response"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupOrder_Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Order/LookupBatch": {
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "Cleansia.Web.Customer.Controllers.OrderController.LookupOrderBatch (Cleansia.Web.Customer)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupOrderBatch_Query"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupOrderBatch_Query"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LookupOrderBatch_Query"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LookupOrderBatch_Response"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupOrderBatch_Response"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupOrderBatch_Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/Order/CreateOrder": {
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "Cleansia.Web.Customer.Controllers.OrderController.CreateOrder (Cleansia.Web.Customer)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder_Command"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder_Command"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder_Command"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrder_Response"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrder_Response"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrder_Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Order/Quote": {
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "Cleansia.Web.Customer.Controllers.OrderController.Quote (Cleansia.Web.Customer)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteOrder_Command"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteOrder_Command"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteOrder_Command"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOrder_Response"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOrder_Response"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOrder_Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Order/GetMyOrders": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "Cleansia.Web.Customer.Controllers.OrderController.GetMyOrders (Cleansia.Web.Customer)",
        "parameters": [
          {
            "name": "Filter.Id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.IsActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Filter.CustomerName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.CustomerEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.CustomerPhone",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.DisplayOrderNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.EmployeeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.CleaningDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Filter.CleaningDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Filter.PaymentStatuses",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PaymentStatus"
              }
            }
          },
          {
            "name": "Filter.PaymentTypes",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PaymentType"
              }
            }
          },
          {
            "name": "Filter.MinTotalPrice",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Filter.MaxTotalPrice",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Filter.OrderStatuses",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/OrderStatus"
              }
            }
          },
          {
            "name": "Filter.HasAvailableSpots",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Filter.IsUnassigned",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Filter.ExcludeEmployeeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SortDefinition"
              }
            }
          },
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "maximum": 500,
              "minimum": 0,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "maximum": 100000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedDataOfOrderListItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedDataOfOrderListItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedDataOfOrderListItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Order/GetPaged": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "Cleansia.Web.Customer.Controllers.OrderController.GetPaged (Cleansia.Web.Customer)",
        "parameters": [
          {
            "name": "Filter.Id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.IsActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Filter.CustomerName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.CustomerEmail",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.CustomerPhone",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.DisplayOrderNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.EmployeeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter.CleaningDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Filter.CleaningDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Filter.PaymentStatuses",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PaymentStatus"
              }
            }
          },
          {
            "name": "Filter.PaymentTypes",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PaymentType"
              }
            }
          },
          {
            "name": "Filter.MinTotalPrice",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Filter.MaxTotalPrice",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Filter.OrderStatuses",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/OrderStatus"
              }
            }
          },
          {
            "name": "Filter.HasAvailableSpots",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Filter.IsUnassigned",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Filter.ExcludeEmployeeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SortDefinition"
              }
            }
          },
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "maximum": 500,
              "minimum": 0,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "maximum": 100000,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedDataOfOrderListItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedDataOfOrderListItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedDataOfOrderListItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Order/GetById": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "Cleansia.Web.Customer.Controllers.OrderController.GetById (Cleansia.Web.Customer)",
        "parameters": [
          {
            "name": "OrderId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LookupOrderBatch_Query": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOrderBatch_OrderLookupItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LookupOrderBatch_Response": {
        "type": "object",
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LookupOrder_Response"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LookupOrder_Response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "displayOrderNumber": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "cleaningDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "paymentType": {
            "$ref": "#/components/schemas/Code"
          },
          "paymentStatus": {
            "$ref": "#/components/schemas/Code"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "estimatedTime": {
            "type": "integer",
            "format": "int32"
          },
          "orderStatus": {
            "$ref": "#/components/schemas/Code"
          },
          "confirmationCode": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDetailDto"
          },
          "selectedServices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceDetails"
            },
            "nullable": true
          },
          "selectedPackages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageDetails"
            },
            "nullable": true
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderStatusTrackDto"
            },
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "PaymentStatus": {
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Pending",
          "Paid",
          "Failed",
          "Refunded",
          "Disputed"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}